doc/netlink: Add specs for addr and route rtnetlink message types
Add support for:
 - newaddr, deladdr, getaddr (dump)
 - getroute (dump)

There is not yet support for notifications because ynl currently doesn't
support defining 'event' properties on a 'do' operation. I plan to work
on this in a future patch.

The link message types are a work in progress that I plan to submit in a
future patchset. Links contain different nested attributes dependent on
the type of link. Decoding these will need some kind of attr-space
selection based on the value of another attribute in the message.

Signed-off-by: Donald Hunter <donald.hunter@gmail.com>