tools/net/ynl: Add a ynl-gen-strace tool
This includes a refactoring of the code from ynl-gen-c.py into several
files in lib/ so that the code can be shared by other code generators.
The ynl-gen-c.py script has not yet been refactored.

This is a proof of concept that would need to be completely redone on
the most recent codebase. Note that the refactoring should really be
spread across several commits.

Signed-off-by: Donald Hunter <donald.hunter@gmail.com>